The lightning ability is now fully functional. However, there is a strange bug with it still hitting enemies that have already been killed. This can be seen when the lightning hits the shrapnel of a fallen enemy.
--HG--
extra : convert_revision : svn%3Acdefcc1b-bc0c-407a-80b0-b0ae7b0029a2%40167

+++ ParticleStorm/player.h
@@ -137,7 +137,7 @@ private:
     // Data members.
     // These represent the location of the player on the last call to update.
     double x_old, y_old;
-    int mana;
+    double mana;
     long int score;
     // The player's currently selected special ability. This should only be
     // assigned values from the ability enum.
